feat: update

+++ server/server.js
@@ -21,17 +21,42 @@ const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 const scriptDir = __dirname;
 const optionsFile = path.join(scriptDir, "rofi-audio.txt");
-const lines = fs
-  .readFileSync(optionsFile, "utf-8")
-  .split("\n")
-  .map((line) => line.trim())
-  .filter((line) => !!line);
-if (!lines.length) throw Error("no lines loaded");
 
-// console.log(lines);
+function loadOptions() {
+  const lines = fs
+    .readFileSync(optionsFile, "utf-8")
+    .split("\n")
+    .map((line) => line.trim())
+    .filter((line) => !!line);
+  if (!lines.length) throw Error("no lines loaded");
+
+  return lines;
+}
+
+function getlineIndex(index) {
+  const lines = loadOptions();
+  return getElemOfNonEmptyArray(lines, index);
+}
+
+// from 1 to inf, if more then lenght - use %
+function getElemOfNonEmptyArray(array, index) {
+  if (index < 0) throw new Error("index must be positive");
+  const length = array.length;
+  if (length === 0) throw new Error("array must not be empty");
+  if (index >= length) return array[length - 1];
+  return array[index];
+}
+
+// getElemOfNonEmptyArray([], 0); // throw error
+// getElemOfNonEmptyArray([1, 2, 3], 0); // 1
+// getElemOfNonEmptyArray([1, 2, 3], 1); // 2
+// getElemOfNonEmptyArray([1, 2, 3], 2); // 3
+// getElemOfNonEmptyArray([1, 2, 3], 3); // 3
+// getElemOfNonEmptyArray([1, "foo", 3], 4); // "foo"
+
 const state = {
   autoplaying: false,
-  lastReadLine: 0,
+  lastReadLineIndex: 0,
   currentAudioProcess: null,
 };
 
@@ -40,11 +65,11 @@ const generateMp3Path = (lineText) => {
   return path.join(process.env.HOME, "Documents/rofi-audio", `${hash}.mp3`);
 };
 
-const playAudio = async (lineNumber) => {
-  console.log("playAudio called with lineNumber:", lineNumber);
-  const lineText = lines[lineNumber - 1];
+const playAudio = async (lineIndex) => {
+  console.log("playAudio called with lineIndex:", lineIndex);
+  const lineText = getlineIndex(lineIndex);
   if (!lineText) {
-    throw new Error(`No text found for line number ${lineNumber}`);
+    throw new Error(`No text found for line number ${lineIndex}`);
   }
 
   const mp3File = generateMp3Path(lineText);
@@ -58,14 +83,18 @@ const playAudio = async (lineNumber) => {
   }
 
   console.log(`Playing: ${lineText}`);
-  await execFile("notify-send", [`${lineNumber}: ${lineText}`]);
+  await execFile("notify-send", [`${lineIndex}: ${lineText}`]);
 
-  state.lastReadLine = lineNumber;
+  state.lastReadLineIndex = lineIndex;
 
   return new Promise((resolve, reject) => {
-    const childProcess = spawn("mplayer", [mp3File], {
-      stdio: ["ignore", process.stdout, process.stdout],
-    });
+    const childProcess = spawn(
+      "mplayer",
+      ["-speed", "1.3", "-af", "scaletempo", mp3File],
+      {
+        stdio: ["ignore", process.stdout, process.stdout],
+      },
+    );
     state.currentAudioProcess = childProcess;
 
     childProcess.on("error", (err) => {
@@ -84,7 +113,7 @@ const playAudio = async (lineNumber) => {
       }
     });
 
-    childProcess.on("exit", (code, signal) => {
+    childProcess.on("exit", (_code, signal) => {
       if (signal === "SIGTERM") {
         resolve();
       }
@@ -98,7 +127,7 @@ app.get("/next", async (_req, res) => {
   if (state.currentAudioProcess) {
     state.currentAudioProcess.kill();
   }
-  await playAudio(state.lastReadLine + 1);
+  await playAudio(state.lastReadLineIndex + 1);
   res.send("Playing next line");
 });
 
@@ -106,7 +135,7 @@ app.get("/prev", async (_req, res) => {
   if (state.currentAudioProcess) {
     state.currentAudioProcess.kill();
   }
-  await playAudio(state.lastReadLine - 1);
+  await playAudio(state.lastReadLineIndex - 1);
   res.send("Playing previous line");
 });
 
@@ -118,22 +147,30 @@ app.get("/stop", (_req, res) => {
   res.send("Stopped audio");
 });
 
-app.get("/autoplay_start", async (_req, res) => {
+app.get("/autoplay_start", async (req, res) => {
   if (!state.autoplaying) {
     state.autoplaying = true;
     while (state.autoplaying) {
       if (state.currentAudioProcess) {
         state.currentAudioProcess.kill();
       }
-      await playAudio(state.lastReadLine + 1);
-      await new Promise((resolve) => setTimeout(resolve, 1000));
+      await new Promise((resolve) =>
+        setTimeout(resolve, req.params.waitMilliseconds || 2000),
+      );
+      if (!state.autoplaying) break; // Check if stopped during wait
+      await playAudio(state.lastReadLineIndex + 1);
+      // await new Promise((resolve) => setTimeout(resolve, 3000));
     }
   }
   res.send("Autoplay started");
 });
 
-app.get("/autoplay_stop", (_req, res) => {
+app.get("/autoplay_stop", async (_req, res) => {
   state.autoplaying = false;
+  if (state.currentAudioProcess) {
+    state.currentAudioProcess.kill();
+  }
+  state.lastReadLineIndex = 0;
   res.send("Autoplay stopped");
 });
 
@@ -142,14 +179,14 @@ app.get("/choose/:line", async (req, res) => {
     state.currentAudioProcess.kill();
   }
   console.log("Playing ", req.params);
-  const lineNumber = parseInt(req.params.line, 10);
-  await playAudio(lineNumber);
-  res.send(`Playing chosen line ${lineNumber}`);
+  const lineIndex = parseInt(req.params.line, 10);
+  await playAudio(lineIndex);
+  res.send(`Playing chosen line ${lineIndex}`);
 });
 
 async function showRofiDialog_() {
   try {
-    const inputText = lines
+    const inputText = loadOptions()
       .map((line, index) => `${index + 1}: ${line}`)
       .join("\n");
     return execFileSync("rofi", ["-dmenu", "-p", "Play"], {
@@ -157,36 +194,39 @@ async function showRofiDialog_() {
       encoding: "utf-8",
     });
   } catch (err) {
-    if (err.code === 1) {
-      throw new Error("No selection made: code 1");
+    if (err.status === 1) {
+      console.log("No selection made: code 1");
+      return null;
     }
     throw err;
   }
-  // const { stdout } = await execFile("rofi", ["-dmenu", "-p", "Play"], {
-  //   encoding: "utf-8",
-  //   input: Buffer.from(inputText),
-  // });
 }
 
 async function showRofiDialog() {
   const stdout = await showRofiDialog_();
+  if (!stdout) {
+    return null;
+  }
   const chosen = stdout.trim();
   if (!chosen) {
     throw new Error("No selection made");
   }
-  const lineNumberString = chosen.split(":")[0].trim();
-  return parseInt(lineNumberString, 10);
+  const lineIndexString = chosen.split(":")[0].trim();
+  return parseInt(lineIndexString, 10);
 }
 
 app.get("/rofi", async (_req, res) => {
-  const lineNumber = await showRofiDialog();
-  console.log("lineNumber", lineNumber);
+  const lineIndex = await showRofiDialog();
+  if (!lineIndex) {
+    return null;
+  }
+  console.log("lineIndex", lineIndex);
   if (state.currentAudioProcess) {
     state.currentAudioProcess.kill();
   }
-  await playAudio(lineNumber);
-  res.send(`Playing rofi selection: ${lineNumber}`);
+  await playAudio(lineIndex);
+  res.send(`Playing rofi selection: ${lineIndex}`);
 });
 
-const port = process.env.PORT || 3000;
+const port = process.env.PORT || 3300;
 app.listen(port, () => console.log(`Server running on port ${port}`));
